+++ cmd/tools/print/tools_print.go
@@ -15,10 +15,15 @@
 package print
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"os"
+	"path"
+	"path/filepath"
+	"regexp"
 	"strconv"
+	"strings"
 
 	"github.com/spf13/cobra"
 	"github.com/streamingfast/bstream"
@@ -27,9 +32,10 @@ import (
 	"github.com/streamingfast/dstore"
 	firecore "github.com/streamingfast/firehose-core"
 	"github.com/streamingfast/firehose-core/types"
+	"go.uber.org/zap"
 )
 
-func NewToolsPrintCmd[B firecore.Block](chain *firecore.Chain[B]) *cobra.Command {
+func NewToolsPrintCmd[B firecore.Block](chain *firecore.Chain[B], logger *zap.Logger) *cobra.Command {
 	toolsPrintCmd := &cobra.Command{
 		Use:   "print",
 		Short: "Prints of one block or merged blocks file",
@@ -42,9 +48,9 @@ func NewToolsPrintCmd[B firecore.Block](chain *firecore.Chain[B]) *cobra.Command
 	}
 
 	toolsPrintMergedBlocksCmd := &cobra.Command{
-		Use:   "merged-blocks <store> <start_block>",
-		Short: "Prints the content summary of a merged blocks file.",
-		Args:  cobra.ExactArgs(2),
+		Use:   "merged-blocks <store|file> [<start_block>[:<end_block>]]",
+		Short: "Prints merged blocks file from store, using range if specified",
+		Args:  cobra.RangeArgs(1, 2),
 	}
 
 	toolsPrintCmd.AddCommand(toolsPrintOneBlockCmd)
@@ -53,63 +59,98 @@ func NewToolsPrintCmd[B firecore.Block](chain *firecore.Chain[B]) *cobra.Command
 	toolsPrintCmd.PersistentFlags().Bool("transactions", false, "When in 'text' output mode, also print transactions summary")
 
 	toolsPrintOneBlockCmd.RunE = createToolsPrintOneBlockE(chain)
-	toolsPrintMergedBlocksCmd.RunE = createToolsPrintMergedBlocksE(chain)
+	toolsPrintMergedBlocksCmd.RunE = createToolsPrintMergedBlocksE(chain, logger)
 
 	return toolsPrintCmd
 }
 
-func createToolsPrintMergedBlocksE[B firecore.Block](chain *firecore.Chain[B]) firecore.CommandExecutor {
+func createToolsPrintMergedBlocksE[B firecore.Block](chain *firecore.Chain[B], logger *zap.Logger) firecore.CommandExecutor {
 	return func(cmd *cobra.Command, args []string) error {
-		ctx := cmd.Context()
-
 		outputPrinter, err := GetOutputPrinter(cmd, chain.BlockFileDescriptor())
 		cli.NoError(err, "Unable to get output printer")
 
-		storeURL := args[0]
-		store, err := dstore.NewDBinStore(storeURL)
-		if err != nil {
-			return fmt.Errorf("unable to create store at path %q: %w", store, err)
-		}
+		store, err := dstore.NewDBinStore(args[0])
+		cli.NoError(err, "Unable to create store %q", args[0])
 
-		startBlock, err := strconv.ParseUint(args[1], 10, 64)
-		if err != nil {
-			return fmt.Errorf("invalid base block %q: %w", args[1], err)
-		}
-		blockBoundary := types.RoundToBundleStartBlock(startBlock, 100)
+		blockRange := types.NewOpenRange(int64(bstream.GetProtocolFirstStreamableBlock))
+		if len(args) > 1 {
+			blockRange, err = types.GetBlockRangeFromArg(args[1])
+			cli.NoError(err, "Unable to parse block range %q", args[1])
 
-		filename := fmt.Sprintf("%010d", blockBoundary)
-		reader, err := store.OpenObject(ctx, filename)
-		if err != nil {
-			fmt.Printf("❌ Unable to read blocks filename %s: %s\n", filename, err)
-			return err
-		}
-		defer reader.Close()
-
-		readerFactory, err := bstream.NewDBinBlockReader(reader)
-		if err != nil {
-			fmt.Printf("❌ Unable to read blocks filename %s: %s\n", filename, err)
-			return err
-		}
-
-		seenBlockCount := 0
-		for {
-			block, err := readerFactory.Read()
-			if err != nil {
-				if err == io.EOF {
-					fmt.Fprintf(os.Stderr, "Total blocks: %d\n", seenBlockCount)
-					return nil
-				}
-				return fmt.Errorf("error receiving blocks: %w", err)
+			// If the range is open, we assume the user wants to print a single block
+			if blockRange.IsOpen() {
+				blockRange = types.NewClosedRange(blockRange.GetStartBlock(), uint64(blockRange.GetStartBlock())+1)
 			}
 
-			seenBlockCount++
+			cli.Ensure(blockRange.IsResolved(), "Invalid block range %q: print only accepts fully resolved range", blockRange)
+		}
 
-			if err := displayBlock(block, chain, outputPrinter); err != nil {
-				// Error is ready to be passed to the user as-is
-				return err
+		if looksLikeStoreURLFile(args[0]) {
+			store, err = dstore.NewDBinStore(storeURLFromFileInput(args[0]))
+			cli.NoError(err, "Unable to create store %q", path.Dir(args[0]))
+
+			storeFileRange := storeURLFileLikeRange(args[0])
+			if len(args) <= 1 {
+				// No block range explicitly specified, we will use the range from the filename
+				blockRange = storeFileRange
+			} else {
+				cli.Ensure(storeFileRange.Contains(uint64(blockRange.GetStartBlock()), types.EndBoundaryExclusive), "Block range %q is not contained in the store file range %q", blockRange, storeFileRange)
+				cli.Ensure(storeFileRange.Contains(uint64(blockRange.MustGetStopBlock()), types.EndBoundaryExclusive), "Block range %q is not contained in the store file range %q", blockRange, storeFileRange)
 			}
 		}
+
+		options := []bstream.FileSourceOption{
+			bstream.FileSourceErrorOnMissingMergedBlocksFile(),
+		}
+		if !blockRange.IsOpen() {
+			options = append(options, bstream.FileSourceWithStopBlock(blockRange.MustGetStopBlock()))
+		}
+
+		source := bstream.NewFileSource(store, uint64(blockRange.GetStartBlock()), bstream.HandlerFunc(func(blk *pbbstream.Block, obj any) error {
+			if !blockRange.Contains(blk.Number, types.RangeBoundaryExclusive) {
+				return nil
+			}
+
+			return displayBlock(blk, chain, outputPrinter)
+		}), logger, options...)
+
+		// Blocking call, perform the work we asked for
+		source.Run()
+
+		if source.Err() != nil && !errors.Is(source.Err(), bstream.ErrStopBlockReached) {
+			return source.Err()
+		}
+
+		return nil
 	}
+}
+
+var mergedBlocksFileRegex = regexp.MustCompile(`^(\d{10})(?:.dbin(?:.zst)?)?$`)
+
+func looksLikeStoreURLFile(path string) bool {
+	base := filepath.Base(path)
+
+	return mergedBlocksFileRegex.MatchString(base)
+}
+
+func storeURLFileLikeRange(path string) types.BlockRange {
+	base := filepath.Base(path)
+	groups := mergedBlocksFileRegex.FindStringSubmatch(base)
+	if len(groups) != 2 {
+		panic(fmt.Sprintf("unexpected number of groups (got %d, expected 2) in regex %q match against %q", len(groups), mergedBlocksFileRegex, base))
+	}
+
+	startBlock, _ := strconv.ParseUint(groups[1], 10, 64)
+	return types.NewClosedRange(int64(startBlock), uint64(startBlock)+100)
+}
+
+func storeURLFromFileInput(input string) string {
+	output := strings.TrimRight(input, filepath.Base(input))
+	output = strings.TrimRightFunc(output, func(r rune) bool {
+		return r == '/' || r == '\\'
+	})
+
+	return output
 }
 
 func createToolsPrintOneBlockE[B firecore.Block](chain *firecore.Chain[B]) firecore.CommandExecutor {
@@ -177,7 +218,7 @@ func displayBlock[B firecore.Block](pbBlock *pbbstream.Block, chain *firecore.Ch
 	}
 
 	if !firecore.UnsafeRunningFromFirecore {
-		// since we are running via the chain specific binary (i.e. fireeth) we can use a BlockFactory
+		// Since we are running via a chain specific binary (i.e. fireeth) we can use a BlockFactory
 		marshallableBlock := chain.BlockFactory()
 
 		if err := pbBlock.Payload.UnmarshalTo(marshallableBlock); err != nil {