+++ cmd/tools/tools.go
@@ -63,7 +63,7 @@ func ConfigureToolsCmd[B firecore.Block](
 	}
 
 	ToolsCmd.AddCommand(check.NewCheckCommand(chain, logger))
-	ToolsCmd.AddCommand(print2.NewToolsPrintCmd(chain))
+	ToolsCmd.AddCommand(print2.NewToolsPrintCmd(chain, logger))
 
 	ToolsCmd.AddCommand(compare.NewToolsCompareBlocksCmd(chain))
 	ToolsCmd.AddCommand(firehose.NewToolsDownloadFromFirehoseCmd(chain, logger))